+++ www/OSFirebaseCloudMessaging.js
@@ -12,20 +12,16 @@ exports.unsubscribe = function (topic, success, error) {
     exec(success, error, 'OSFirebaseCloudMessaging', 'unsubscribe', [topic]);
 };
 
-exports.requestPermission = function ( success, error) {
-    exec(success, error, 'OSFirebaseCloudMessaging', 'requestPermission');
-};
-
 exports.clearNotifications = function (success, error) {
     exec(success, error, 'OSFirebaseCloudMessaging', 'clearNotifications');
 };
 
 exports.setBadge = function (badge, success, error) {
-    exec(success, error, 'OSFirebaseCloudMessaging', 'setBadge', [badge, title, body]);
+    exec(success, error, 'OSFirebaseCloudMessaging', 'setBadge', [badge]);
 };
 
-exports.sendLocalNotification = function (badge, title, body, success, error) {
-    exec(success, error, 'OSFirebaseCloudMessaging', 'sendLocalNotification', [badge, title, body]);
+exports.sendLocalNotification = function (badge, title, body, channelName, channelDescription, success, error) {
+    exec(success, error, 'OSFirebaseCloudMessaging', 'sendLocalNotification', [badge, title, body, channelName, channelDescription]);
 };
 
 exports.getBadge = function (success, error) {
@@ -38,8 +34,4 @@ exports.registerDevice = function (success, error) {
 
 exports.unregisterDevice = function (success, error) {
     exec(success, error, 'OSFirebaseCloudMessaging', 'unregisterDevice');
-};
-
-exports.sendLocalNotification = function (badge, title, body, success, error) {
-    exec(success, error, 'OSFirebaseCloudMessaging', 'sendLocalNotification', [badge, title, body]);
 };